Revert "Merge pull request #101 from wafflestudio19-5/refactor/tag"
This reverts commit 023acd4ff119ecb9f4e41dd73c4964cbe1510ee8, reversing
changes made to 6249e5e3b747bad3bda201c7f9f914c73b01f6dd.
