+++ project/config/urls.py
@@ -26,7 +26,6 @@ urlpatterns = [
     path("admin/", admin.site.urls),
     path("api/v1/", include("user.urls")),
     path("api/v1/", include("newsfeed.urls")),
-    path("api/v1/", include("notice.urls")),
     path(
         "swagger/",
         schema_view.with_ui("swagger", cache_timeout=0),
