+++ project/newsfeed/models.py
@@ -2,7 +2,6 @@ from django.db import models
 from django.db.models.deletion import CASCADE
 from django.db.models.fields.related import ForeignKey
 from user.models import User
-from .utils import get_directory_path, comment_directory_path
 
 
 class NewsfeedObject(models.Model):
@@ -15,6 +14,16 @@ class NewsfeedObject(models.Model):
 
     class Meta:
         abstract = True
+
+
+def get_directory_path(instance, filename):
+    return (
+        f"user/{instance.author}/posts/{instance.mainpost.id}/{instance.id}/{filename}"
+    )
+
+
+def comment_directory_path(instance, filename):
+    return f"user/{instance.author}/comments/{instance.id}/{filename}"
 
 
 class Post(NewsfeedObject):
@@ -49,8 +58,6 @@ class Post(NewsfeedObject):
         User, blank=True, related_name="notice_off_posts"
     )
 
-    tagged_users = models.ManyToManyField(User, blank=True, related_name="tagged_posts")
-
     def get_user_url(self):
         # 게시글에서 유저를 누르면 유저 프로필로 갈 수 있게 하기 위함
         return f"/api/v1/user/{self.author}/"
@@ -68,9 +75,29 @@ class Comment(NewsfeedObject):
 
     file = models.FileField(upload_to=comment_directory_path, blank=True)
 
-    tagged_users = models.ManyToManyField(
-        User, blank=True, related_name="tagged_comments"
-    )
-
     def get_user_url(self):
         return f"/api/v1/user/{self.author}/"
+
+
+class Notice(models.Model):
+
+    id = models.AutoField(primary_key=True)
+    user = models.ForeignKey(User, on_delete=CASCADE, related_name="notices")
+    post = models.ForeignKey(Post, on_delete=CASCADE, null=True, related_name="notices")
+    parent_comment = models.ForeignKey(
+        Comment, on_delete=CASCADE, null=True, related_name="notices"
+    )
+    content = models.CharField(max_length=30, blank=True)
+    created = models.DateTimeField(auto_now=True)
+    is_checked = models.BooleanField(default=False)
+    url = models.CharField(max_length=1000)
+
+
+class NoticeSender(models.Model):
+    notice = models.ForeignKey(
+        Notice, on_delete=CASCADE, null=True, related_name="senders"
+    )
+    user = models.ForeignKey(
+        User, on_delete=CASCADE, null=True, related_name="noticesenders"
+    )
+    count = models.PositiveIntegerField(default=0)
