+++ project/newsfeed/tests.py
@@ -4,7 +4,10 @@ from factory.django import DjangoModelFactory
 from faker import Faker
 from user.models import User
 from newsfeed.models import Post, Comment
+from django.test import TestCase
 from rest_framework import status
+import json
+from datetime import datetime
 from django.core.files.uploadedfile import SimpleUploadedFile
 import os
 from pathlib import Path
@@ -66,6 +69,958 @@ class CommentFactory(DjangoModelFactory):
         model = Comment
 
 
+class NoticeTestCase(TestCase):
+    @classmethod
+    def setUpTestData(cls):
+        cls.test_user = UserFactory.create(
+            email="notice@test.com",
+            password="password",
+            first_name="재현",
+            last_name="이",
+            birth="1997-02-03",
+            gender="M",
+            phone_number="01000000000",
+        )
+        cls.test_friends = UserFactory.create_batch(10)
+        cls.user_token = "JWT " + jwt_token_of(cls.test_user)
+        cls.friends_token = []
+        for friend in cls.test_friends:
+            cls.test_user.friends.add(friend)
+            cls.friends_token.append("JWT " + jwt_token_of(friend))
+
+        cls.test_post = PostFactory.create(
+            author=cls.test_user, content="알림 테스트 게시물입니다.", likes=0
+        )
+        cls.test_comment = CommentFactory.create(
+            author=cls.test_user, post=cls.test_post, depth=0, content="알림 테스트 댓글입니다."
+        )
+
+    def test_notice(self):
+
+        tmp_comment_list = []
+
+        # 깊이 1인 답글 알림
+        for i, friend_token in enumerate(self.friends_token):
+            response = self.client.post(
+                f"/api/v1/newsfeed/{self.test_post.id}/comment/",
+                data={"content": f"알림 테스트 답글입니다...{i}", "parent": self.test_comment.id},
+                HTTP_AUTHORIZATION=friend_token,
+            )
+            tmp_comment_list.append(response.json()["id"])
+            self.assertEqual(response.status_code, status.HTTP_201_CREATED)
+
+        response = self.client.get(
+            "/api/v1/newsfeed/notices/",
+            content_type="application/json",
+            HTTP_AUTHORIZATION=self.user_token,
+        )
+        self.assertEqual(response.status_code, status.HTTP_200_OK)
+        data = response.json()
+        self.assertEqual(data["results"][0]["content"], "CommentComment")
+        self.assertEqual(data["results"][0]["count"], 9)
+        self.assertEqual(data["results"][0]["post"]["id"], self.test_post.id)
+        self.assertEqual(
+            data["results"][0]["parent_comment"]["comment_id"], self.test_comment.id
+        )
+        self.assertEqual(len(data["results"][0]["senders"]), 9)
+        self.assertEqual(
+            data["results"][0]["sender_preview"]["user"]["id"], self.test_friends[-1].id
+        )
+        self.assertEqual(
+            data["results"][0]["url"],
+            f"api/v1/newsfeed/{self.test_post.id}/{self.test_comment.id}/",
+        )
+        self.assertEqual(data["results"][0]["is_checked"], False)
+        self.assertEqual(
+            data["results"][0]["sender_preview"]["content"], "알림 테스트 답글입니다...9"
+        )
+
+        tmp_comment_id = data["results"][0]["sender_preview"]["comment_id"]
+
+        # 깊이 2인 답글 알림
+        response = self.client.post(
+            f"/api/v1/newsfeed/{self.test_post.id}/comment/",
+            data={"content": "알림 테스트 답글의 답글입니다.", "parent": tmp_comment_id},
+            HTTP_AUTHORIZATION=self.user_token,
+        )
+        self.assertEqual(response.status_code, status.HTTP_201_CREATED)
+
+        response = self.client.get(
+            "/api/v1/newsfeed/notices/",
+            content_type="application/json",
+            HTTP_AUTHORIZATION=self.friends_token[-1],
+        )
+        self.assertEqual(response.status_code, status.HTTP_200_OK)
+        data = response.json()
+        self.assertEqual(data["results"][0]["content"], "CommentComment")
+        self.assertEqual(data["results"][0]["count"], 0)
+        self.assertEqual(data["results"][0]["post"]["id"], self.test_post.id)
+        self.assertEqual(
+            data["results"][0]["parent_comment"]["comment_id"], tmp_comment_id
+        )
+        self.assertEqual(len(data["results"][0]["senders"]), 0)
+        self.assertEqual(
+            data["results"][0]["sender_preview"]["user"]["id"], self.test_user.id
+        )
+        self.assertEqual(
+            data["results"][0]["url"],
+            f"api/v1/newsfeed/{self.test_post.id}/{tmp_comment_id}/",
+        )
+        self.assertEqual(
+            data["results"][0]["sender_preview"]["content"], "알림 테스트 답글의 답글입니다."
+        )
+        tmp_comment_id = data["results"][0]["sender_preview"]["comment_id"]
+
+        # 깊이 2인 답글 삭제 --> 알림 취소
+        response = self.client.delete(
+            f"/api/v1/newsfeed/{self.test_post.id}/{tmp_comment_id}/",
+            HTTP_AUTHORIZATION=self.user_token,
+        )
+        self.assertEqual(response.status_code, status.HTTP_204_NO_CONTENT)
+        response = self.client.get(
+            "/api/v1/newsfeed/notices/",
+            content_type="application/json",
+            HTTP_AUTHORIZATION=self.friends_token[-1],
+        )
+        self.assertEqual(response.status_code, status.HTTP_200_OK)
+        data = response.json()
+        self.assertEqual(len(data["results"]), 0)
+
+        # 깊이 1인 답글들 삭제 --> 알림 취소
+        for i, friend_token in enumerate(self.friends_token):
+            tmp_comment_id = tmp_comment_list[i]
+            response = self.client.delete(
+                f"/api/v1/newsfeed/{self.test_post.id}/{tmp_comment_id}/",
+                HTTP_AUTHORIZATION=friend_token,
+            )
+            self.assertEqual(response.status_code, status.HTTP_204_NO_CONTENT)
+
+        response = self.client.get(
+            "/api/v1/newsfeed/notices/",
+            content_type="application/json",
+            HTTP_AUTHORIZATION=self.user_token,
+        )
+        self.assertEqual(response.status_code, status.HTTP_200_OK)
+        data = response.json()
+
+        self.assertEqual(len(data["results"]), 0)
+
+        # 댓글알림
+        for i, friend_token in enumerate(self.friends_token):
+            response = self.client.post(
+                f"/api/v1/newsfeed/{self.test_post.id}/comment/",
+                data={"content": f"알림 테스트 댓글입니다...{i}"},
+                HTTP_AUTHORIZATION=friend_token,
+            )
+            self.assertEqual(response.status_code, status.HTTP_201_CREATED)
+
+        response = self.client.get(
+            "/api/v1/newsfeed/notices/",
+            content_type="application/json",
+            HTTP_AUTHORIZATION=self.user_token,
+        )
+        self.assertEqual(response.status_code, status.HTTP_200_OK)
+        data = response.json()
+        notice_id = data["results"][0]["id"]
+
+        self.assertEqual(len(data["results"]), 1)
+        self.assertEqual(data["results"][0]["content"], "PostComment")
+        self.assertEqual(data["results"][0]["count"], 9)
+        self.assertEqual(data["results"][0]["post"]["id"], self.test_post.id)
+        self.assertEqual(len(data["results"][0]["senders"]), 9)
+        self.assertEqual(
+            data["results"][0]["sender_preview"]["user"]["id"], self.test_friends[-1].id
+        )
+        self.assertEqual(
+            data["results"][0]["url"], f"api/v1/newsfeed/{self.test_post.id}/"
+        )
+        self.assertEqual(data["results"][0]["is_checked"], False)
+        self.assertEqual(
+            data["results"][0]["sender_preview"]["content"], "알림 테스트 댓글입니다...9"
+        )
+
+        # 댓글을 단 사람이 또 다른 댓글을 다는 경우
+        response = self.client.post(
+            f"/api/v1/newsfeed/{self.test_post.id}/comment/",
+            data={"content": "이미 댓글을 단 사람은, 또 댓글을 달아도 count가 늘어나지 않습니다."},
+            HTTP_AUTHORIZATION=self.friends_token[0],
+        )
+        self.assertEqual(response.status_code, status.HTTP_201_CREATED)
+
+        response = self.client.get(
+            f"/api/v1/newsfeed/notices/",
+            content_type="application/json",
+            HTTP_AUTHORIZATION=self.user_token,
+        )
+        self.assertEqual(response.status_code, status.HTTP_200_OK)
+        data = response.json()
+        self.assertEqual(data["results"][0]["count"], 9)
+        self.assertEqual(len(data["results"][0]["senders"]), 9)
+        self.assertEqual(
+            data["results"][0]["sender_preview"]["user"]["id"], self.test_friends[0].id
+        )
+        self.assertEqual(
+            data["results"][0]["sender_preview"]["content"],
+            "이미 댓글을 단 사람은, 또 댓글을 달아도 count가 늘어나지 않습니다.",
+        )
+        recent_comment_id = data["results"][0]["sender_preview"]["comment_id"]
+
+        # 알림 is_checked
+        response = self.client.get(
+            f"/api/v1/newsfeed/notices/{notice_id}/",
+            content_type="application/json",
+            HTTP_AUTHORIZATION=self.user_token,
+        )
+        self.assertEqual(response.status_code, status.HTTP_200_OK)
+        data = response.json()
+        self.assertEqual(data["is_checked"], True)
+
+        # 자기자신 알림 X
+        response = self.client.post(
+            f"/api/v1/newsfeed/{self.test_post.id}/comment/",
+            data={"content": "본인이 단 댓글은 알림에 뜨지 않습니다."},
+            HTTP_AUTHORIZATION=self.user_token,
+        )
+        self.assertEqual(response.status_code, status.HTTP_201_CREATED)
+
+        response = self.client.get(
+            f"/api/v1/newsfeed/notices/",
+            content_type="application/json",
+            HTTP_AUTHORIZATION=self.user_token,
+        )
+        self.assertEqual(response.status_code, status.HTTP_200_OK)
+        data = response.json()
+        self.assertEqual(data["results"][0]["count"], 9)
+        self.assertEqual(len(data["results"][0]["senders"]), 9)
+
+        response = self.client.post(
+            f"/api/v1/newsfeed/{self.test_post.id}/comment/",
+            data={"content": "본인이 단 답글은 알림에 뜨지 않습니다.", "parent": recent_comment_id},
+            HTTP_AUTHORIZATION=self.friends_token[0],
+        )
+        self.assertEqual(response.status_code, status.HTTP_201_CREATED)
+
+        response = self.client.get(
+            f"/api/v1/newsfeed/notices/",
+            content_type="application/json",
+            HTTP_AUTHORIZATION=self.friends_token[0],
+        )
+        self.assertEqual(response.status_code, status.HTTP_200_OK)
+        data = response.json()
+        self.assertEqual(len(data["results"]), 0)
+
+        # 댓글을 2개 이상 단 사람이, 하나를 삭제해도 count는 그대로 !
+        response = self.client.delete(
+            f"/api/v1/newsfeed/{self.test_post.id}/{recent_comment_id}/",
+            HTTP_AUTHORIZATION=self.friends_token[0],
+        )
+        self.assertEqual(response.status_code, status.HTTP_204_NO_CONTENT)
+
+        response = self.client.get(
+            f"/api/v1/newsfeed/notices/",
+            content_type="application/json",
+            HTTP_AUTHORIZATION=self.user_token,
+        )
+        self.assertEqual(response.status_code, status.HTTP_200_OK)
+        data = response.json()
+        self.assertEqual(data["results"][0]["count"], 9)
+        self.assertEqual(len(data["results"][0]["senders"]), 9)
+        self.assertEqual(
+            data["results"][0]["sender_preview"]["user"]["id"], self.test_friends[-1].id
+        )
+        self.assertEqual(
+            data["results"][0]["sender_preview"]["content"], "알림 테스트 댓글입니다...9"
+        )
+        recent_comment_id = data["results"][0]["sender_preview"]["comment_id"]
+
+        # 댓글 좋아요
+        for friend_token in self.friends_token:
+            response = self.client.put(
+                f"/api/v1/newsfeed/{self.test_post.id}/{self.test_comment.id}/like/",
+                content_type="application/json",
+                HTTP_AUTHORIZATION=friend_token,
+            )
+        response = self.client.get(
+            "/api/v1/newsfeed/notices/",
+            content_type="application/json",
+            HTTP_AUTHORIZATION=self.user_token,
+        )
+        self.assertEqual(response.status_code, status.HTTP_200_OK)
+        data = response.json()
+        self.assertEqual(len(data["results"]), 2)
+        self.assertEqual(data["results"][0]["content"], "CommentLike")
+        self.assertEqual(data["results"][0]["count"], 9)
+        self.assertEqual(data["results"][0]["post"]["id"], self.test_post.id)
+        self.assertEqual(
+            data["results"][0]["parent_comment"]["comment_id"], self.test_comment.id
+        )
+        self.assertEqual(len(data["results"][0]["senders"]), 9)
+        self.assertEqual(
+            data["results"][0]["sender_preview"]["user"]["id"], self.test_friends[-1].id
+        )
+        self.assertEqual(
+            data["results"][0]["url"],
+            f"api/v1/newsfeed/{self.test_post.id}/",
+        )
+        self.assertEqual(data["results"][0]["is_checked"], False)
+        self.assertEqual(
+            data["results"][0]["parent_comment"]["content"], "알림 테스트 댓글입니다."
+        )
+
+        # 게시글 좋아요
+        for friend_token in self.friends_token:
+            response = self.client.put(
+                f"/api/v1/newsfeed/{self.test_post.id}/like/",
+                content_type="application/json",
+                HTTP_AUTHORIZATION=friend_token,
+            )
+        response = self.client.get(
+            "/api/v1/newsfeed/notices/",
+            content_type="application/json",
+            HTTP_AUTHORIZATION=self.user_token,
+        )
+        self.assertEqual(response.status_code, status.HTTP_200_OK)
+        data = response.json()
+        self.assertEqual(len(data["results"]), 3)
+        notice_id_2 = data["results"][0]["id"]
+        self.assertEqual(data["results"][0]["content"], "PostLike")
+        self.assertEqual(data["results"][0]["count"], 9)
+        self.assertEqual(data["results"][0]["post"]["id"], self.test_post.id)
+        self.assertEqual(len(data["results"][0]["senders"]), 9)
+        self.assertEqual(
+            data["results"][0]["sender_preview"]["user"]["id"], self.test_friends[-1].id
+        )
+        self.assertEqual(
+            data["results"][0]["url"],
+            f"api/v1/newsfeed/{self.test_post.id}/",
+        )
+        self.assertEqual(data["results"][0]["is_checked"], False)
+
+        # 알림 취소
+        response = self.client.put(
+            f"/api/v1/newsfeed/{self.test_post.id}/like/",
+            content_type="application/json",
+            HTTP_AUTHORIZATION=self.friends_token[0],
+        )
+        response = self.client.get(
+            "/api/v1/newsfeed/notices/",
+            content_type="application/json",
+            HTTP_AUTHORIZATION=self.user_token,
+        )
+        self.assertEqual(response.status_code, status.HTTP_200_OK)
+        data = response.json()
+        self.assertEqual(data["results"][0]["count"], 8)
+        self.assertEqual(len(data["results"][0]["senders"]), 8)
+        self.assertEqual(
+            data["results"][0]["sender_preview"]["user"]["id"], self.test_friends[-1].id
+        )
+
+        response = self.client.put(
+            f"/api/v1/newsfeed/{self.test_post.id}/{self.test_comment.id}/like/",
+            content_type="application/json",
+            HTTP_AUTHORIZATION=self.friends_token[0],
+        )
+        response = self.client.get(
+            "/api/v1/newsfeed/notices/",
+            content_type="application/json",
+            HTTP_AUTHORIZATION=self.user_token,
+        )
+        self.assertEqual(response.status_code, status.HTTP_200_OK)
+        data = response.json()
+        self.assertEqual(data["results"][1]["count"], 8)
+        self.assertEqual(len(data["results"][1]["senders"]), 8)
+        self.assertEqual(
+            data["results"][1]["sender_preview"]["user"]["id"], self.test_friends[-1].id
+        )
+
+        response = self.client.delete(
+            f"/api/v1/newsfeed/{self.test_post.id}/{recent_comment_id}/",
+            HTTP_AUTHORIZATION=self.friends_token[9],
+        )
+        self.assertEqual(response.status_code, status.HTTP_204_NO_CONTENT)
+        response = self.client.get(
+            "/api/v1/newsfeed/notices/",
+            content_type="application/json",
+            HTTP_AUTHORIZATION=self.user_token,
+        )
+        self.assertEqual(response.status_code, status.HTTP_200_OK)
+        data = response.json()
+        self.assertEqual(data["results"][2]["count"], 8)
+        self.assertEqual(len(data["results"][2]["senders"]), 8)
+        self.assertEqual(
+            data["results"][2]["sender_preview"]["content"], "알림 테스트 댓글입니다...8"
+        )
+        self.assertEqual(
+            data["results"][2]["sender_preview"]["user"]["id"], self.test_friends[-2].id
+        )
+
+        # 알림 삭제
+        response = self.client.delete(
+            f"/api/v1/newsfeed/notices/{notice_id}/",
+            content_type="application/json",
+            HTTP_AUTHORIZATION=self.user_token,
+        )
+        self.assertEqual(response.status_code, status.HTTP_204_NO_CONTENT)
+
+        response = self.client.get(
+            "/api/v1/newsfeed/notices/",
+            content_type="application/json",
+            HTTP_AUTHORIZATION=self.user_token,
+        )
+        self.assertEqual(response.status_code, status.HTTP_200_OK)
+        data = response.json()
+        self.assertEqual(len(data["results"]), 2)
+        self.assertNotEqual(data["results"][-1]["id"], notice_id)
+
+        # 댓글 알림 삭제한 후에 댓글이 달려 알림이 오는 경우
+        for i, friend_token in enumerate(self.friends_token):
+            response = self.client.post(
+                f"/api/v1/newsfeed/{self.test_post.id}/comment/",
+                data={"content": f"새로운 알림 테스트 댓글입니다...{i}"},
+                HTTP_AUTHORIZATION=friend_token,
+            )
+            self.assertEqual(response.status_code, status.HTTP_201_CREATED)
+
+        response = self.client.get(
+            "/api/v1/newsfeed/notices/",
+            content_type="application/json",
+            HTTP_AUTHORIZATION=self.user_token,
+        )
+        self.assertEqual(response.status_code, status.HTTP_200_OK)
+        data = response.json()
+
+        self.assertEqual(len(data["results"]), 3)
+        self.assertEqual(data["results"][0]["content"], "PostComment")
+        self.assertEqual(data["results"][0]["count"], 9)
+        self.assertEqual(data["results"][0]["post"]["id"], self.test_post.id)
+        self.assertEqual(len(data["results"][0]["senders"]), 9)
+        self.assertEqual(
+            data["results"][0]["sender_preview"]["user"]["id"], self.test_friends[-1].id
+        )
+        self.assertEqual(
+            data["results"][0]["url"], f"api/v1/newsfeed/{self.test_post.id}/"
+        )
+        self.assertEqual(data["results"][0]["is_checked"], False)
+        self.assertEqual(
+            data["results"][0]["sender_preview"]["content"], "새로운 알림 테스트 댓글입니다...9"
+        )
+
+        # 좋아요 알림 삭제한 후에 좋아요 알림이 새로 오는 경우
+        response = self.client.delete(
+            f"/api/v1/newsfeed/notices/{notice_id_2}/",
+            content_type="application/json",
+            HTTP_AUTHORIZATION=self.user_token,
+        )
+        self.assertEqual(response.status_code, status.HTTP_204_NO_CONTENT)
+
+        for i in range(1, 10):
+            response = self.client.put(
+                f"/api/v1/newsfeed/{self.test_post.id}/like/",
+                content_type="application/json",
+                HTTP_AUTHORIZATION=self.friends_token[i],
+            )
+        for friend_token in self.friends_token:
+            response = self.client.put(
+                f"/api/v1/newsfeed/{self.test_post.id}/like/",
+                content_type="application/json",
+                HTTP_AUTHORIZATION=friend_token,
+            )
+        response = self.client.get(
+            "/api/v1/newsfeed/notices/",
+            content_type="application/json",
+            HTTP_AUTHORIZATION=self.user_token,
+        )
+        self.assertEqual(response.status_code, status.HTTP_200_OK)
+        data = response.json()
+        self.assertEqual(len(data["results"]), 3)
+        self.assertEqual(data["results"][0]["content"], "PostLike")
+        self.assertEqual(data["results"][0]["count"], 9)
+        self.assertEqual(data["results"][0]["post"]["id"], self.test_post.id)
+        self.assertEqual(len(data["results"][0]["senders"]), 9)
+        self.assertEqual(
+            data["results"][0]["sender_preview"]["user"]["id"], self.test_friends[-1].id
+        )
+        self.assertEqual(
+            data["results"][0]["url"],
+            f"api/v1/newsfeed/{self.test_post.id}/",
+        )
+        self.assertEqual(data["results"][0]["is_checked"], False)
+
+        # 친구요청
+        stranger = UserFactory.create()
+        stranger_token = "JWT " + jwt_token_of(stranger)
+
+        response = self.client.post(
+            f"/api/v1/friend/request/{self.test_user.id}/",
+            content_type="application/json",
+            HTTP_AUTHORIZATION=stranger_token,
+        )
+        self.assertEqual(response.status_code, status.HTTP_201_CREATED)
+        response = self.client.get(
+            "/api/v1/newsfeed/notices/",
+            content_type="application/json",
+            HTTP_AUTHORIZATION=self.user_token,
+        )
+        self.assertEqual(response.status_code, status.HTTP_200_OK)
+        data = response.json()
+        self.assertEqual(data["results"][0]["content"], "FriendRequest")
+        self.assertEqual(
+            data["results"][0]["sender_preview"]["user"]["id"], stranger.id
+        )
+        self.assertEqual(len(data["results"][0]["senders"]), 0)
+
+        # 친구수락
+        response = self.client.put(
+            f"/api/v1/friend/request/{stranger.id}/",
+            data={"sender": stranger.id},
+            content_type="application/json",
+            HTTP_AUTHORIZATION=self.user_token,
+        )
+        self.assertEqual(response.status_code, status.HTTP_200_OK)
+
+        response = self.client.get(
+            "/api/v1/newsfeed/notices/",
+            content_type="application/json",
+            HTTP_AUTHORIZATION=self.user_token,
+        )
+        self.assertEqual(response.status_code, status.HTTP_200_OK)
+        data = response.json()
+        self.assertEqual(data["results"][0]["content"], "isFriend")
+        self.assertEqual(data["results"][0]["is_checked"], True)
+        self.assertEqual(
+            data["results"][0]["sender_preview"]["user"]["id"], stranger.id
+        )
+        self.assertEqual(len(data["results"][0]["senders"]), 0)
+
+        response = self.client.get(
+            "/api/v1/newsfeed/notices/",
+            content_type="application/json",
+            HTTP_AUTHORIZATION=stranger_token,
+        )
+        self.assertEqual(response.status_code, status.HTTP_200_OK)
+        data = response.json()
+        self.assertEqual(data["results"][0]["content"], "FriendAccept")
+        self.assertEqual(
+            data["results"][0]["sender_preview"]["user"]["id"], self.test_user.id
+        )
+        self.assertEqual(len(data["results"][0]["senders"]), 0)
+
+        # 친구 요청 후, 요청한 사람이 친구 요청을 취소할 시, 보내졌던 알림 삭제
+        stranger2 = UserFactory.create()
+        stranger2_token = "JWT " + jwt_token_of(stranger2)
+
+        response = self.client.post(
+            f"/api/v1/friend/request/{self.test_user.id}/",
+            content_type="application/json",
+            HTTP_AUTHORIZATION=stranger2_token,
+        )
+        self.assertEqual(response.status_code, status.HTTP_201_CREATED)
+
+        response = self.client.get(
+            "/api/v1/newsfeed/notices/",
+            content_type="application/json",
+            HTTP_AUTHORIZATION=self.user_token,
+        )
+        self.assertEqual(response.status_code, status.HTTP_200_OK)
+        data = response.json()
+        self.assertEqual(data["results"][0]["content"], "FriendRequest")
+        self.assertEqual(
+            data["results"][0]["sender_preview"]["user"]["id"], stranger2.id
+        )
+        self.assertEqual(len(data["results"][0]["senders"]), 0)
+
+        response = self.client.delete(
+            f"/api/v1/friend/request/{self.test_user.id}/",
+            content_type="application/json",
+            HTTP_AUTHORIZATION=stranger2_token,
+        )
+        self.assertEqual(response.status_code, status.HTTP_204_NO_CONTENT)
+        response = self.client.get(
+            "/api/v1/newsfeed/notices/",
+            content_type="application/json",
+            HTTP_AUTHORIZATION=self.user_token,
+        )
+        self.assertEqual(response.status_code, status.HTTP_200_OK)
+        data = response.json()
+        self.assertNotEqual(
+            data["results"][0]["sender_preview"]["user"]["id"], stranger2.id
+        )
+
+        # 친구 요청 후, 요청 받은 사람의 친구 요청 삭제시 알림도 삭제
+        response = self.client.post(
+            f"/api/v1/friend/request/{self.test_user.id}/",
+            content_type="application/json",
+            HTTP_AUTHORIZATION=stranger2_token,
+        )
+        self.assertEqual(response.status_code, status.HTTP_201_CREATED)
+
+        response = self.client.get(
+            "/api/v1/newsfeed/notices/",
+            content_type="application/json",
+            HTTP_AUTHORIZATION=self.user_token,
+        )
+        self.assertEqual(response.status_code, status.HTTP_200_OK)
+        data = response.json()
+        self.assertEqual(data["results"][0]["content"], "FriendRequest")
+        self.assertEqual(
+            data["results"][0]["sender_preview"]["user"]["id"], stranger2.id
+        )
+        self.assertEqual(len(data["results"][0]["senders"]), 0)
+
+        response = self.client.delete(
+            f"/api/v1/friend/request/{stranger2.id}/",
+            content_type="application/json",
+            HTTP_AUTHORIZATION=self.user_token,
+        )
+        self.assertEqual(response.status_code, status.HTTP_204_NO_CONTENT)
+        response = self.client.get(
+            "/api/v1/newsfeed/notices/",
+            content_type="application/json",
+            HTTP_AUTHORIZATION=self.user_token,
+        )
+        self.assertEqual(response.status_code, status.HTTP_200_OK)
+        data = response.json()
+        self.assertNotEqual(
+            data["results"][0]["sender_preview"]["user"]["id"], stranger2.id
+        )
+
+    def test_notice_on_off(self):
+
+        # 게시물 작성자가 게시물에 대한 알림 끄기
+        response = self.client.put(
+            f"/api/v1/newsfeed/{self.test_post.id}/notice/",
+            content_type="application/json",
+            HTTP_AUTHORIZATION=self.user_token,
+        )
+        self.assertEqual(response.status_code, status.HTTP_200_OK)
+        data = response.json()
+        self.assertEqual(data["is_noticed"], False)
+
+        # 댓글 작성, 알림 X
+        response = self.client.post(
+            f"/api/v1/newsfeed/{self.test_post.id}/comment/",
+            data={"content": "알림이 발생하지 않습니다."},
+            HTTP_AUTHORIZATION=self.friends_token[0],
+        )
+        self.assertEqual(response.status_code, status.HTTP_201_CREATED)
+        data = response.json()
+        comment_id = data["id"]
+
+        response = self.client.get(
+            "/api/v1/newsfeed/notices/",
+            content_type="application/json",
+            HTTP_AUTHORIZATION=self.user_token,
+        )
+        self.assertEqual(response.status_code, status.HTTP_200_OK)
+        data = response.json()
+        self.assertEqual(len(data["results"]), 0)
+
+        # 게시글 좋아요, 알림 X
+        response = self.client.put(
+            f"/api/v1/newsfeed/{self.test_post.id}/like/",
+            content_type="application/json",
+            HTTP_AUTHORIZATION=self.friends_token[0],
+        )
+        self.assertEqual(response.status_code, status.HTTP_200_OK)
+
+        response = self.client.get(
+            "/api/v1/newsfeed/notices/",
+            content_type="application/json",
+            HTTP_AUTHORIZATION=self.user_token,
+        )
+        self.assertEqual(response.status_code, status.HTTP_200_OK)
+        data = response.json()
+        self.assertEqual(len(data["results"]), 0)
+
+        # 게시글 좋아요 취소, 에러 발생 X
+        response = self.client.put(
+            f"/api/v1/newsfeed/{self.test_post.id}/like/",
+            content_type="application/json",
+            HTTP_AUTHORIZATION=self.friends_token[0],
+        )
+
+        # 게시글 댓글 삭제, 에러 발생 X
+        response = self.client.delete(
+            f"/api/v1/newsfeed/{self.test_post.id}/{comment_id}/",
+            HTTP_AUTHORIZATION=self.friends_token[0],
+        )
+        self.assertEqual(response.status_code, status.HTTP_204_NO_CONTENT)
+
+        # 게시물에 대한 알림 켜기
+        response = self.client.put(
+            f"/api/v1/newsfeed/{self.test_post.id}/notice/",
+            content_type="application/json",
+            HTTP_AUTHORIZATION=self.user_token,
+        )
+        self.assertEqual(response.status_code, status.HTTP_200_OK)
+        data = response.json()
+        self.assertEqual(data["is_noticed"], True)
+
+        # 댓글 작성, 알림 O
+        test_image = SimpleUploadedFile(
+            name="testimage2.jpg",
+            content=open(os.path.join(BASE_DIR, "testimage2.jpg"), "rb").read(),
+            content_type="image/jpeg",
+        )
+
+        response = self.client.post(
+            f"/api/v1/newsfeed/{self.test_post.id}/comment/",
+            data={"content": "알림이 발생합니다.", "file": test_image},
+            HTTP_AUTHORIZATION=self.friends_token[0],
+        )
+        self.assertEqual(response.status_code, status.HTTP_201_CREATED)
+        data = response.json()
+        comment_id = data["id"]
+
+        response = self.client.get(
+            "/api/v1/newsfeed/notices/",
+            content_type="application/json",
+            HTTP_AUTHORIZATION=self.user_token,
+        )
+        self.assertEqual(response.status_code, status.HTTP_200_OK)
+        data = response.json()
+        self.assertEqual(len(data["results"]), 1)
+        self.assertEqual(data["results"][0]["content"], "PostComment")
+        self.assertEqual(data["results"][0]["count"], 0)
+        self.assertEqual(data["results"][0]["post"]["id"], self.test_post.id)
+        self.assertEqual(len(data["results"][0]["senders"]), 0)
+        self.assertEqual(
+            data["results"][0]["sender_preview"]["user"]["id"], self.test_friends[0].id
+        )
+        self.assertEqual(
+            data["results"][0]["url"], f"api/v1/newsfeed/{self.test_post.id}/"
+        )
+        self.assertEqual(data["results"][0]["is_checked"], False)
+        self.assertEqual(data["results"][0]["sender_preview"]["content"], "알림이 발생합니다.")
+        self.assertIn("photo", data["results"][0]["sender_preview"]["is_file"])
+
+        # 게시글 좋아요, 알림 O
+        response = self.client.put(
+            f"/api/v1/newsfeed/{self.test_post.id}/like/",
+            content_type="application/json",
+            HTTP_AUTHORIZATION=self.friends_token[0],
+        )
+        self.assertEqual(response.status_code, status.HTTP_200_OK)
+
+        response = self.client.get(
+            "/api/v1/newsfeed/notices/",
+            content_type="application/json",
+            HTTP_AUTHORIZATION=self.user_token,
+        )
+        self.assertEqual(response.status_code, status.HTTP_200_OK)
+        data = response.json()
+        self.assertEqual(len(data["results"]), 2)
+        self.assertEqual(data["results"][0]["content"], "PostLike")
+        self.assertEqual(data["results"][0]["count"], 0)
+        self.assertEqual(data["results"][0]["post"]["id"], self.test_post.id)
+        self.assertEqual(len(data["results"][0]["senders"]), 0)
+        self.assertEqual(
+            data["results"][0]["sender_preview"]["user"]["id"], self.test_friends[0].id
+        )
+        self.assertEqual(
+            data["results"][0]["url"],
+            f"api/v1/newsfeed/{self.test_post.id}/",
+        )
+        self.assertEqual(data["results"][0]["is_checked"], False)
+
+        # 댓글 작성자의 게시물에 대한 알림 끄기
+        response = self.client.put(
+            f"/api/v1/newsfeed/{self.test_post.id}/notice/",
+            content_type="application/json",
+            HTTP_AUTHORIZATION=self.friends_token[0],
+        )
+        self.assertEqual(response.status_code, status.HTTP_200_OK)
+        data = response.json()
+        self.assertEqual(data["is_noticed"], False)
+
+        # 댓글 좋아요 해도 알림 X
+        response = self.client.put(
+            f"/api/v1/newsfeed/{self.test_post.id}/{comment_id}/like/",
+            content_type="application/json",
+            HTTP_AUTHORIZATION=self.friends_token[1],
+        )
+
+        response = self.client.get(
+            "/api/v1/newsfeed/notices/",
+            content_type="application/json",
+            HTTP_AUTHORIZATION=self.friends_token[0],
+        )
+        self.assertEqual(response.status_code, status.HTTP_200_OK)
+        data = response.json()
+        self.assertEqual(len(data["results"]), 0)
+
+        # 답글 달기, 알림 X
+        response = self.client.post(
+            f"/api/v1/newsfeed/{self.test_post.id}/comment/",
+            data={"content": f"알림이 발생하지 않는 답글입니다.", "parent": comment_id},
+            HTTP_AUTHORIZATION=self.friends_token[2],
+        )
+        self.assertEqual(response.status_code, status.HTTP_201_CREATED)
+
+        response = self.client.get(
+            "/api/v1/newsfeed/notices/",
+            content_type="application/json",
+            HTTP_AUTHORIZATION=self.friends_token[0],
+        )
+        self.assertEqual(response.status_code, status.HTTP_200_OK)
+        data = response.json()
+        self.assertEqual(len(data["results"]), 0)
+
+        # 댓글 작성자의 게시물에 대한 알림 켜기
+        response = self.client.put(
+            f"/api/v1/newsfeed/{self.test_post.id}/notice/",
+            content_type="application/json",
+            HTTP_AUTHORIZATION=self.friends_token[0],
+        )
+        self.assertEqual(response.status_code, status.HTTP_200_OK)
+        data = response.json()
+        self.assertEqual(data["is_noticed"], True)
+
+        # 댓글 좋아요, 알림 O
+        response = self.client.put(
+            f"/api/v1/newsfeed/{self.test_post.id}/{comment_id}/like/",
+            content_type="application/json",
+            HTTP_AUTHORIZATION=self.friends_token[1],
+        )
+
+        response = self.client.put(
+            f"/api/v1/newsfeed/{self.test_post.id}/{comment_id}/like/",
+            content_type="application/json",
+            HTTP_AUTHORIZATION=self.friends_token[1],
+        )
+
+        response = self.client.get(
+            "/api/v1/newsfeed/notices/",
+            content_type="application/json",
+            HTTP_AUTHORIZATION=self.friends_token[0],
+        )
+        self.assertEqual(response.status_code, status.HTTP_200_OK)
+        data = response.json()
+        self.assertEqual(len(data["results"]), 1)
+        self.assertEqual(data["results"][0]["content"], "CommentLike")
+        self.assertEqual(data["results"][0]["count"], 0)
+        self.assertEqual(data["results"][0]["post"]["id"], self.test_post.id)
+        self.assertEqual(data["results"][0]["parent_comment"]["comment_id"], comment_id)
+        self.assertEqual(len(data["results"][0]["senders"]), 0)
+        self.assertEqual(
+            data["results"][0]["sender_preview"]["user"]["id"], self.test_friends[1].id
+        )
+        self.assertEqual(
+            data["results"][0]["url"],
+            f"api/v1/newsfeed/{self.test_post.id}/",
+        )
+        self.assertEqual(data["results"][0]["is_checked"], False)
+        self.assertEqual(data["results"][0]["parent_comment"]["content"], "알림이 발생합니다.")
+
+        # 답글 달기, 알림 O
+        response = self.client.post(
+            f"/api/v1/newsfeed/{self.test_post.id}/comment/",
+            data={"content": "알림이 발생하는 답글입니다.", "parent": comment_id},
+            HTTP_AUTHORIZATION=self.friends_token[3],
+        )
+        self.assertEqual(response.status_code, status.HTTP_201_CREATED)
+
+        response = self.client.get(
+            "/api/v1/newsfeed/notices/",
+            content_type="application/json",
+            HTTP_AUTHORIZATION=self.friends_token[0],
+        )
+        self.assertEqual(response.status_code, status.HTTP_200_OK)
+        data = response.json()
+        self.assertEqual(data["results"][0]["content"], "CommentComment")
+        self.assertEqual(data["results"][0]["count"], 0)
+        self.assertEqual(data["results"][0]["post"]["id"], self.test_post.id)
+        self.assertEqual(data["results"][0]["parent_comment"]["comment_id"], comment_id)
+        self.assertEqual(len(data["results"][0]["senders"]), 0)
+        self.assertEqual(
+            data["results"][0]["sender_preview"]["user"]["id"], self.test_friends[3].id
+        )
+        self.assertEqual(
+            data["results"][0]["url"],
+            f"api/v1/newsfeed/{self.test_post.id}/{comment_id}/",
+        )
+        self.assertEqual(data["results"][0]["is_checked"], False)
+        self.assertEqual(
+            data["results"][0]["sender_preview"]["content"], "알림이 발생하는 답글입니다."
+        )
+
+        # subpost 단위로 알림 꺼보기
+        data = {
+            "content": "메인 포스트입니다.",
+            "subposts": ["첫번째 포스트입니다."],
+            "file": [test_image],
+        }
+
+        response = self.client.post(
+            "/api/v1/newsfeed/",
+            data=data,
+            HTTP_AUTHORIZATION=self.user_token,
+        )
+        self.assertEqual(response.status_code, status.HTTP_201_CREATED)
+        subpost_id = response.json()["subposts"][0]["id"]
+
+        response = self.client.put(
+            f"/api/v1/newsfeed/{subpost_id}/notice/",
+            content_type="application/json",
+            HTTP_AUTHORIZATION=self.user_token,
+        )
+        self.assertEqual(response.status_code, status.HTTP_200_OK)
+        data = response.json()
+        self.assertEqual(data["is_noticed"], False)
+
+        response = self.client.post(
+            f"/api/v1/newsfeed/{subpost_id}/comment/",
+            data={"content": "subpost 알림이 발생하지 않습니다."},
+            HTTP_AUTHORIZATION=self.friends_token[0],
+        )
+        self.assertEqual(response.status_code, status.HTTP_201_CREATED)
+
+        response = self.client.get(
+            "/api/v1/newsfeed/notices/",
+            content_type="application/json",
+            HTTP_AUTHORIZATION=self.user_token,
+        )
+        self.assertEqual(response.status_code, status.HTTP_200_OK)
+        data = response.json()
+        self.assertNotEqual(data["results"][0]["content"], "PostComment")
+
+        # subpost 단위로 알림 받기
+        response = self.client.put(
+            f"/api/v1/newsfeed/{subpost_id}/notice/",
+            content_type="application/json",
+            HTTP_AUTHORIZATION=self.user_token,
+        )
+        self.assertEqual(response.status_code, status.HTTP_200_OK)
+        data = response.json()
+        self.assertEqual(data["is_noticed"], True)
+
+        response = self.client.post(
+            f"/api/v1/newsfeed/{subpost_id}/comment/",
+            data={"content": "subpost 알림이 발생합니다."},
+            HTTP_AUTHORIZATION=self.friends_token[1],
+        )
+        self.assertEqual(response.status_code, status.HTTP_201_CREATED)
+
+        response = self.client.get(
+            "/api/v1/newsfeed/notices/",
+            content_type="application/json",
+            HTTP_AUTHORIZATION=self.user_token,
+        )
+        self.assertEqual(response.status_code, status.HTTP_200_OK)
+        data = response.json()
+
+        self.assertEqual(data["results"][0]["content"], "PostComment")
+        self.assertEqual(data["results"][0]["count"], 0)
+        self.assertEqual(data["results"][0]["post"]["id"], subpost_id)
+        self.assertEqual(len(data["results"][0]["senders"]), 0)
+        self.assertEqual(
+            data["results"][0]["sender_preview"]["user"]["id"], self.test_friends[1].id
+        )
+        self.assertEqual(
+            data["results"][0]["sender_preview"]["content"], "subpost 알림이 발생합니다."
+        )
+
+
 class NewsFeedTestCase(TestCase):
     @classmethod
     def setUpTestData(cls):
@@ -117,7 +1072,6 @@ class NewsFeedTestCase(TestCase):
         PostFactory.create(
             author=cls.test_stranger, content="모르는 사람의 테스트 게시물입니다.", likes=30
         )
-        cls.content_type = "multipart/form-data; boundary=BoUnDaRyStRiNg"
 
     def test_post_list(self):
 
@@ -223,8 +1177,7 @@ class NewsFeedTestCase(TestCase):
 
         response = self.client.post(
             "/api/v1/newsfeed/",
-            data=encode_multipart("BoUnDaRyStRiNg", data),
-            content_type=self.content_type,
+            data=data,
             HTTP_AUTHORIZATION=user_token,
         )
         self.assertEqual(response.status_code, status.HTTP_201_CREATED)
@@ -257,8 +1210,7 @@ class NewsFeedTestCase(TestCase):
         }
         response = self.client.post(
             "/api/v1/newsfeed/",
-            data=encode_multipart("BoUnDaRyStRiNg", data),
-            content_type=self.content_type,
+            data=data,
             HTTP_AUTHORIZATION=user_token,
         )
         self.assertEqual(response.status_code, status.HTTP_201_CREATED)
@@ -268,8 +1220,7 @@ class NewsFeedTestCase(TestCase):
         }
         response = self.client.post(
             "/api/v1/newsfeed/",
-            data=encode_multipart("BoUnDaRyStRiNg", data),
-            content_type=self.content_type,
+            data=data,
             HTTP_AUTHORIZATION=user_token,
         )
         self.assertEqual(response.status_code, status.HTTP_201_CREATED)
@@ -280,8 +1231,7 @@ class NewsFeedTestCase(TestCase):
         }
         response = self.client.post(
             "/api/v1/newsfeed/",
-            data=encode_multipart("BoUnDaRyStRiNg", data),
-            content_type=self.content_type,
+            data=data,
             HTTP_AUTHORIZATION=user_token,
         )
         self.assertEqual(response.status_code, status.HTTP_201_CREATED)
@@ -289,8 +1239,8 @@ class NewsFeedTestCase(TestCase):
         data = {}
         response = self.client.post(
             "/api/v1/newsfeed/",
-            data=encode_multipart("BoUnDaRyStRiNg", data),
-            content_type=self.content_type,
+            data=data,
+            content_type="application/json",
             HTTP_AUTHORIZATION=user_token,
         )
         self.assertEqual(response.status_code, status.HTTP_400_BAD_REQUEST)
@@ -304,8 +1254,7 @@ class NewsFeedTestCase(TestCase):
         }
         response = self.client.post(
             "/api/v1/newsfeed/",
-            data=encode_multipart("BoUnDaRyStRiNg", data),
-            content_type=self.content_type,
+            data=data,
             HTTP_AUTHORIZATION=user_token,
         )
         self.assertEqual(response.status_code, status.HTTP_201_CREATED)
@@ -355,8 +1304,7 @@ class NewsFeedTestCase(TestCase):
 
         response = self.client.post(
             "/api/v1/newsfeed/",
-            data=encode_multipart("BoUnDaRyStRiNg", data),
-            content_type=self.content_type,
+            data=data,
             HTTP_AUTHORIZATION=user_token,
         )
         self.assertEqual(response.status_code, status.HTTP_201_CREATED)
@@ -613,7 +1561,6 @@ class ShareTestCase(TestCase):
         cls.user_token = "JWT " + jwt_token_of(cls.test_user)
         cls.friend_token = "JWT " + jwt_token_of(cls.test_friend)
         cls.stranger_token = "JWT " + jwt_token_of(cls.test_stranger)
-        cls.content_type = "multipart/form-data; boundary=BoUnDaRyStRiNg"
 
     def test_post_share(self):
 
@@ -626,8 +1573,7 @@ class ShareTestCase(TestCase):
 
         response = self.client.post(
             "/api/v1/newsfeed/",
-            data=encode_multipart("BoUnDaRyStRiNg", data),
-            content_type=self.content_type,
+            data=data,
             HTTP_AUTHORIZATION=self.user_token,
         )
         self.assertEqual(response.status_code, status.HTTP_201_CREATED)
@@ -640,8 +1586,7 @@ class ShareTestCase(TestCase):
 
         response = self.client.post(
             "/api/v1/newsfeed/",
-            data=encode_multipart("BoUnDaRyStRiNg", data),
-            content_type=self.content_type,
+            data=data,
             HTTP_AUTHORIZATION=self.user_token,
         )
         self.assertEqual(response.status_code, status.HTTP_201_CREATED)
@@ -703,7 +1648,7 @@ class ShareTestCase(TestCase):
         self.assertEqual(response.status_code, status.HTTP_200_OK)
         data = response.json()
         self.assertEqual(data["results"][0]["content"], "게시글을 공유했습니다.")
-        self.assertEqual(data["results"][0]["shared_post"], None)
+        self.assertEqual(data["results"][0]["shared_post"], "AccessDenied")
 
         data = {
             "content": "",
@@ -729,7 +1674,7 @@ class ShareTestCase(TestCase):
         self.assertEqual(response.status_code, status.HTTP_200_OK)
         data = response.json()
         self.assertEqual(data["results"][0]["content"], "게시글을 공유했습니다.")
-        self.assertEqual(data["results"][0]["shared_post"], None)
+        self.assertEqual(data["results"][0]["shared_post"], "AccessDenied")
 
         response = self.client.get(
             f"/api/v1/user/{self.test_user.id}/newsfeed/",
@@ -749,7 +1694,7 @@ class ShareTestCase(TestCase):
         self.assertEqual(response.status_code, status.HTTP_200_OK)
         data = response.json()
         self.assertEqual(data["results"][0]["content"], "게시글을 공유했습니다.")
-        self.assertEqual(data["results"][0]["shared_post"], None)
+        self.assertEqual(data["results"][0]["shared_post"], "AccessDenied")
 
         # subpost도 공유가 가능한지
         test_image = SimpleUploadedFile(
@@ -765,8 +1710,7 @@ class ShareTestCase(TestCase):
 
         response = self.client.post(
             "/api/v1/newsfeed/",
-            data=encode_multipart("BoUnDaRyStRiNg", data),
-            content_type=self.content_type,
+            data=data,
             HTTP_AUTHORIZATION=self.user_token,
         )
         self.assertEqual(response.status_code, status.HTTP_201_CREATED)
@@ -775,8 +1719,7 @@ class ShareTestCase(TestCase):
 
         response = self.client.post(
             "/api/v1/newsfeed/",
-            data=encode_multipart("BoUnDaRyStRiNg", data),
-            content_type=self.content_type,
+            data=data,
             HTTP_AUTHORIZATION=self.user_token,
         )
         self.assertEqual(response.status_code, status.HTTP_201_CREATED)
@@ -839,7 +1782,6 @@ class ScopeTestCase(TestCase):
         cls.user_token = "JWT " + jwt_token_of(cls.test_user)
         cls.friend_token = "JWT " + jwt_token_of(cls.test_friend)
         cls.stranger_token = "JWT " + jwt_token_of(cls.test_stranger)
-        cls.content_type = "multipart/form-data; boundary=BoUnDaRyStRiNg"
 
     def test_scope_list(self):
 
@@ -850,8 +1792,7 @@ class ScopeTestCase(TestCase):
 
         response = self.client.post(
             "/api/v1/newsfeed/",
-            data=encode_multipart("BoUnDaRyStRiNg", data),
-            content_type=self.content_type,
+            data=data,
             HTTP_AUTHORIZATION=self.user_token,
         )
         self.assertEqual(response.status_code, status.HTTP_201_CREATED)
@@ -911,8 +1852,7 @@ class ScopeTestCase(TestCase):
 
         response = self.client.post(
             "/api/v1/newsfeed/",
-            data=encode_multipart("BoUnDaRyStRiNg", data),
-            content_type=self.content_type,
+            data=data,
             HTTP_AUTHORIZATION=self.user_token,
         )
         self.assertEqual(response.status_code, status.HTTP_201_CREATED)
@@ -972,8 +1912,7 @@ class ScopeTestCase(TestCase):
 
         response = self.client.post(
             "/api/v1/newsfeed/",
-            data=encode_multipart("BoUnDaRyStRiNg", data),
-            content_type=self.content_type,
+            data=data,
             HTTP_AUTHORIZATION=self.user_token,
         )
         self.assertEqual(response.status_code, status.HTTP_201_CREATED)
@@ -1041,8 +1980,7 @@ class ScopeTestCase(TestCase):
 
         response = self.client.post(
             "/api/v1/newsfeed/",
-            data=encode_multipart("BoUnDaRyStRiNg", data),
-            content_type=self.content_type,
+            data=data,
             HTTP_AUTHORIZATION=self.user_token,
         )
         self.assertEqual(response.status_code, status.HTTP_201_CREATED)
@@ -1061,8 +1999,7 @@ class ScopeTestCase(TestCase):
 
         response = self.client.post(
             "/api/v1/newsfeed/",
-            data=encode_multipart("BoUnDaRyStRiNg", data),
-            content_type=self.content_type,
+            data=data,
             HTTP_AUTHORIZATION=self.user_token,
         )
         self.assertEqual(response.status_code, status.HTTP_201_CREATED)
@@ -1081,8 +2018,7 @@ class ScopeTestCase(TestCase):
         }
         response = self.client.post(
             "/api/v1/newsfeed/",
-            data=encode_multipart("BoUnDaRyStRiNg", data),
-            content_type=self.content_type,
+            data=data,
             HTTP_AUTHORIZATION=self.user_token,
         )
         self.assertEqual(response.status_code, status.HTTP_400_BAD_REQUEST)
@@ -1095,8 +2031,7 @@ class ScopeTestCase(TestCase):
         }
         response = self.client.post(
             "/api/v1/newsfeed/",
-            data=encode_multipart("BoUnDaRyStRiNg", data),
-            content_type=self.content_type,
+            data=data,
             HTTP_AUTHORIZATION=self.user_token,
         )
         self.assertEqual(response.status_code, status.HTTP_400_BAD_REQUEST)
@@ -1109,8 +2044,7 @@ class ScopeTestCase(TestCase):
         }
         response = self.client.post(
             "/api/v1/newsfeed/",
-            data=encode_multipart("BoUnDaRyStRiNg", data),
-            content_type=self.content_type,
+            data=data,
             HTTP_AUTHORIZATION=self.user_token,
         )
         self.assertEqual(response.status_code, status.HTTP_400_BAD_REQUEST)
@@ -1126,8 +2060,7 @@ class ScopeTestCase(TestCase):
 
         response = self.client.post(
             "/api/v1/newsfeed/",
-            data=encode_multipart("BoUnDaRyStRiNg", data),
-            content_type=self.content_type,
+            data=data,
             HTTP_AUTHORIZATION=self.user_token,
         )
         self.assertEqual(response.status_code, status.HTTP_201_CREATED)
@@ -1222,8 +2155,7 @@ class ScopeTestCase(TestCase):
 
         response = self.client.post(
             f"/api/v1/newsfeed/{mainpost_id}/comment/",
-            data=encode_multipart("BoUnDaRyStRiNg", {"content": "전체공개 게시글의 댓글입니다."}),
-            content_type=self.content_type,
+            data={"content": "전체공개 게시글의 댓글입니다."},
             HTTP_AUTHORIZATION=self.stranger_token,
         )
         self.assertEqual(response.status_code, status.HTTP_201_CREATED)
@@ -1350,7 +2282,6 @@ class LikeTestCase(TestCase):
         for user in cls.users:
             cls.test_friend.friends.add(user)
         cls.test_friend.save()
-        cls.content_type = "multipart/form-data; boundary=BoUnDaRyStRiNg"
 
     def test_like_and_unlike(self):  # 좋아요하고 해제하기
         user = self.test_user
@@ -1518,7 +2449,6 @@ class CommentTestCase(TestCase):
         CommentFactory.create_batch(
             5, author=cls.test_friend, post=cls.my_post, depth=2, parent=cls.depth_one
         )
-        cls.content_type = "multipart/form-data; boundary=BoUnDaRyStRiNg"
 
     def test_comment_list(self):
 
@@ -1575,8 +2505,7 @@ class CommentTestCase(TestCase):
 
         response = self.client.post(
             f"/api/v1/newsfeed/{self.friend_post.id}/comment/",
-            data=encode_multipart("BoUnDaRyStRiNg", data),
-            content_type=self.content_type,
+            data=data,
             HTTP_AUTHORIZATION=user_token,
         )
         self.assertEqual(response.status_code, status.HTTP_201_CREATED)
@@ -1596,8 +2525,7 @@ class CommentTestCase(TestCase):
 
         response = self.client.post(
             f"/api/v1/newsfeed/{self.friend_post.id}/comment/",
-            data=encode_multipart("BoUnDaRyStRiNg", data),
-            content_type=self.content_type,
+            data=data,
             HTTP_AUTHORIZATION=user_token,
         )
 
@@ -1612,8 +2540,8 @@ class CommentTestCase(TestCase):
         data = {}
         response = self.client.post(
             f"/api/v1/newsfeed/{self.my_post.id}/comment/",
-            data=encode_multipart("BoUnDaRyStRiNg", data),
-            content_type=self.content_type,
+            data=data,
+            content_type="application/json",
             HTTP_AUTHORIZATION=user_token,
         )
         self.assertEqual(response.status_code, status.HTTP_400_BAD_REQUEST)
@@ -1626,8 +2554,8 @@ class CommentTestCase(TestCase):
         }
         response = self.client.post(
             f"/api/v1/newsfeed/{self.my_post.id}/comment/",
-            data=encode_multipart("BoUnDaRyStRiNg", data),
-            content_type=self.content_type,
+            data=data,
+            content_type="application/json",
             HTTP_AUTHORIZATION=user_token,
         )
         self.assertEqual(response.status_code, status.HTTP_400_BAD_REQUEST)
@@ -1638,7 +2566,7 @@ class CommentTestCase(TestCase):
         }
         response = self.client.post(
             f"/api/v1/newsfeed/{self.stranger_post.id}/comment/",
-            data=encode_multipart("BoUnDaRyStRiNg", data),
+            data=data,
             content_type="application/json",
             HTTP_AUTHORIZATION=user_token,
         )
@@ -1648,8 +2576,8 @@ class CommentTestCase(TestCase):
         data = {"content": content, "parent": self.depth_two.id}
         response = self.client.post(
             f"/api/v1/newsfeed/{self.my_post.id}/comment/",
-            data=encode_multipart("BoUnDaRyStRiNg", data),
-            content_type=self.content_type,
+            data=data,
+            content_type="application/json",
             HTTP_AUTHORIZATION=user_token,
         )
         self.assertEqual(response.status_code, status.HTTP_400_BAD_REQUEST)
@@ -1702,8 +2630,8 @@ class CommentTestCase(TestCase):
 
         response = self.client.put(
             f"/api/v1/newsfeed/{self.my_post.id}/{self.depth_zero.id}/",
-            data=encode_multipart("BoUnDaRyStRiNg", data),
-            content_type=self.content_type,
+            data=data,
+            content_type="application/json",
             HTTP_AUTHORIZATION=friend_token,
         )
         self.assertEqual(response.status_code, status.HTTP_200_OK)
