+++ project/newsfeed/urls.py
@@ -1,14 +1,18 @@
-from django.urls import path
+from django.urls import path, include
 from .views import (
     PostListView,
     PostUpdateView,
     PostLikeView,
     CommentListView,
     CommentLikeView,
+    NoticeView,
+    NoticeListView,
     CommentUpdateDeleteView,
-    TestView,
+    NoticeOnOffView,
 )
-from notice.views import NoticeOnOffView
+from rest_framework.routers import SimpleRouter
+from django.conf import settings
+from django.conf.urls.static import static
 
 urlpatterns = [
     path("newsfeed/", PostListView.as_view()),
@@ -18,5 +22,6 @@ urlpatterns = [
     path("newsfeed/<int:post_id>/<int:comment_id>/", CommentUpdateDeleteView.as_view()),
     path("newsfeed/<int:post_id>/like/", PostLikeView.as_view()),
     path("newsfeed/<int:post_id>/<int:comment_id>/like/", CommentLikeView.as_view()),
-    path("test/", TestView.as_view()),
+    path("newsfeed/notices/", NoticeListView.as_view()),
+    path("newsfeed/notices/<int:notice_id>/", NoticeView.as_view()),
 ]
